[Forget] Fetch less data
Reduced the data usage of the forget handler in two ways.

1. Stop relying on `get_message_content` to retrieve the content
of a message. As the message is guaranteed to already be on
the node, we can directly get the parts of interest from the DB.

2. Only retrieve the fields relevant to the handling of the forget
message from the DB. Rather than retrieving the whole content,
which can be quite large, we now only retrieve the address, item
type and item hash from the DB. This should improve the performance
of the handler. Introduced a new class, `TargetMessageInfo`, to
represent the data we fetch from the DB.