[Messages] Use raw message model everywhere
Propagated the use of the `RawBaseStorage` subclasses in all
functions related to message processing.

Split the message validation function, check_message, in two parts:
* `parse_message` checks that the message is semantically valid,
  i.e. that all the required fields are present, are of the correct
  type and have sensible values.
* `verify_signature` checks the signature of the message using
  the public key of the sender.

We now call parse_message as early as possible in the process.
This allows to simplify hypotheses about the presence/absence
of specific fields and allows to simplify the codebase.

One of the current issues is that the content of the message
can only be loaded conditionally (if the message has inline content).
Otherwise, we must load the content from the network. As this
operation goes beyond simple validation, we perform it later
in the flow, leading to a situation where the content of a message
is not guaranteed to be available at a given point depending
on the `item_type` field of the message.

Modified tests to use the raw message classes. Removed a few tests
that tested corner cases linked to incomplete message dictionaries.