[DB] Fix for capped collection bulk inserts
Implemented a fix (a hack, really) to avoid exceptions when
processing a message several times in the same batch in
the message processing job.

We now isolate operations on capped collections and execute them
one by one. We proceed to ignore errors, as they mean that
the message was already inserted in the capped collection.