Internal: reduce cyclical dependencies in chains code
Problem: the code to interact with chains relied on several global
variables, which ultimately led to cyclical imports and weird side
effects (ex: requiring unneeded imports in tests for the tests to
work).

Solution: introduce classes for the different bricks of message
and signature validation.

The code for each chain is now managed by a `ChainConnector` class
that implements the different operations:

* signature verification (all the chains)
* tx fetch (ETH and NULS2)
* message packing (ETH and NULS2).

All these connectors are owned by a generic `ChainService` class
that provides an abstraction for code that does not care about
the details of which chain does what.

The code in `chains.common` was also split in several files to
reduce dependencies. Message processing is now the responsibility
of the new `MessageHandler` class, which is now defined out of
the chains code.

Most of the functionalities of the pending message job are now
encapsulated in the `PendingMessageProcessor` class.

Some of these changes may seem overkill, but they are required
to get rid of globals entirely in further PRs.