Internal: make storage code more modular
Problem: the code to interact with the local file storage relied
on several global variables, which ultimately obfuscate the code.

Solution: introduce classes to manage the access to the local
storage.

The new `StorageService` class now manages the functions of
the `storage` module. It uses a `StorageEngine` object to interact
(=read/write/delete) with GridFS. Added more classes as needed
for the refactoring.