Fix: increase timeout on pending TX chain data file fetch
Problem: fetching chain data files while processing pending TXs
may time out.

Solution: increase the timeout from 10 to 60 seconds.

We expect chain data files to always be available, so there is
no clear definition of what should be done if the operation really
fails. For the moment, we accept this situation, just increase
the timeout to reduce the amount of exceptions logged (10 seconds
was too short) and let the handler retry, forever if needed.