Feature: publish all processed messages on RabbitMQ
Problem: We wish to give visibility on the message processing status
of all messages from the API.

Solution: extend the capabilities of the Aleph messages MQ topic.

The topic is now a full-blown topic and not just a fanout.
The routing key uses the `<status>.<item_hash>` format.

The new message processing result classes status enum allow to
determine the status of each message in output of the message
processing job.