Feature: POST /messages/ endpoint
Problem: the current way to send messages is:
1. too permissive, as it allows to publish on any P2P channel
2. only allows for async posting.

Solution: introduce a new endpoint, POST /messages/.

This endpoint is the new default for the SDKs and applications
publishing Aleph messages. It supports two modes of publishing:
sync and async. In async mode, the behaviour is similar to
/p2p/pubsub/pub/ as the message is validated, shared with
the other nodes and prepared for processing. The main difference
is that the endpoint will then return with a 202 instead of a 200
code.

In sync mode, the endpoint sets up a connection to the internal
Aleph message MQ and waits for the processing of the message.
It then returns:

* a 200 if the message was successfully processed
* a 422 if an error occurred
* a 202 if the processing of the message took longer than expected.